Second Commit

+++ Assignment3/src/main/java/Assignment3/Assignment3/Callimplementation.java
@@ -17,7 +17,7 @@ public class Callimplementation {
 		
 		return calI.sub(a, b);
 	}
-	public int DivTwoNum(int a,int b)
+	public double DivTwoNum(int a,int b)
 	{
 		return calI.div(a, b);
